refactor(ui): decouple productivity screen into viewmodel and sub-widgets
Extracted business logic, filtering, and data preparation from
`productivity.tsx` into a new `useProductivityViewModel` hook.
Decomposed the 850+ line "God Widget" by moving `renderItem`,
`ListHeaderComponent`, and `ListEmptyComponent` into private named
sub-components. This adheres to Clean Architecture and SRP.
